requirements.txt
google-generativeai
textract
mysql-connector-python